loader and sceneinventory as defaults
