--- avalon/tools/cbloader/__init__.py
@@ -1,10 +1,0 @@
-
-from .app import (
-    show,
-    cli
-)
-
-__all__ = [
-    "show",
-    "cli",
-]
