+++ avalon/tools/loader/__init__.py
@@ -1,7 +1,10 @@
+
 from .app import (
     show,
+    cli
 )
 
 __all__ = [
     "show",
+    "cli",
 ]
