--- avalon/tools/manager/__init__.py
@@ -1,7 +1,0 @@
-from .app import (
-    show,
-)
-
-__all__ = [
-    "show",
-]
