+++ res/houdini/MainMenuCommon.XML
@@ -29,16 +29,16 @@ creator.show()
             <scriptItem id="avalon_load">
                 <label>Load ..</label>
                 <scriptCode><![CDATA[
-from avalon.tools import cbloader
-cbloader.show(use_context=True)
+from avalon.tools import loader
+loader.show(use_context=True)
 ]]></scriptCode>
             </scriptItem>
 
             <scriptItem id="avalon_manage">
                 <label>Manage ..</label>
                 <scriptCode><![CDATA[
-from avalon.tools import cbsceneinventory
-cbsceneinventory.show()
+from avalon.tools import sceneinventory
+sceneinventory.show()
 ]]></scriptCode>
             </scriptItem>
